requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
azure-cosmos==4.5.1
pydantic==2.5.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
python-dotenv==1.0.0
dnspython==2.4.2
tenacity==8.2.3
httpx==0.26.0
